+++ tests/test_search.py
@@ -2,6 +2,7 @@
 
 # pylint: disable=redefined-outer-name
 # pylint: disable=protected-access
+# pylint: disable=import-outside-toplevel
 
 from unittest.mock import MagicMock, patch